+++ src/game.js
@@ -1,8 +1,8 @@
 const data = require('./data');
 const prototypeQuestions = data.prototypeData;
 const util = require('./util');
-const {createDeck, countCards} = require('./card');
-const {createRound} = require('./round');
+const { createDeck, countCards } = require('./card');
+const { createRound } = require('./round');
 
 function printMessage(deck) {
   console.log(`Welcome to FlashCards! You are playing with ${countCards(deck)} cards.
